Refactor parser core into modular parsers
