+++ src/pcap_tool/parser/__init__.py
@@ -3,9 +3,9 @@ from .core import (
     parse_pcap_to_df,
     iter_parsed_frames,
     validate_pcap_file,
-    _safe_int,
     ParserNotAvailable,
 )
+from ..parsers.utils import _safe_int
 
 __all__ = [
     "parse_pcap",
