+++ src/pcap_tool/parser/helpers.py
@@ -1,3 +1,4 @@
-from .core import validate_pcap_file, _safe_int
+from .core import validate_pcap_file
+from ..parsers.utils import _safe_int
 
 __all__ = ["validate_pcap_file", "_safe_int"]
