+++ src/pcap_tool/parser/pcapkit_parser.py
@@ -1,3 +1,3 @@
-from .core import _parse_with_pcapkit, _USE_PCAPKIT as USE_PCAPKIT
+from ..parsers.pcapkit_parser import _parse_with_pcapkit, USE_PCAPKIT
 
 __all__ = ["_parse_with_pcapkit", "USE_PCAPKIT"]
