+++ src/pcap_tool/parser/pyshark_parser.py
@@ -1,3 +1,3 @@
-from .core import _parse_with_pyshark, _USE_PYSHARK as USE_PYSHARK
+from ..parsers.pyshark_parser import _parse_with_pyshark, USE_PYSHARK
 
 __all__ = ["_parse_with_pyshark", "USE_PYSHARK"]
